OBDX-7 Refactored product creation endpoint (#2885)
* OBDX-21 T3. Create endpoint for order creation messages

* changes as per PR change request

* added url mapping for subscribe orders

* Fixed bug on toJson. and handling mappings with null also.

* OBDX-7 Optimized code for create Product and using product service

* OBDX-7 refactoring as Justin comments

* OBDX-7 refactoring as Justin comments
